Moved files into new folders to organise them better and added comments to index to section off article and poster proofing text
