+++ src/components/constants/index.tsx
@@ -1,9 +1,13 @@
 export const proofingText = `There was nothing so very remarkable in that; nor did Alice think it so very much out \n of the way to hear the Rabbit say to itself, “Oh dear! Oh dear! I shall be late!” (when she thought it over afterwards, it occurred to her that she ought to have wondered at this, but at the time it all seemed quite natural); but when the Rabbit actually took a watch out of its waistcoat-pocket, and looked at it, and then hurried on, Alice started to her feet, for it flashed across her mind that she had never before seen a rabbit with either a waistcoat-pocket, or a watch to take out of it, and burning with curiosity, she ran across the field after it, and fortunately was just in time to see it pop down a large rabbit-hole under the hedge.`;
 export const opentypeText = `The quick brown fox jumps over the lazy dog. 1234567890 ABCDEFGHIJKLMNOPQRSTUVWXYZ abcdefghijklmnopqrstuvwxyz !@#$%^*()_+-=[]{}|;':",.<>/?\\`;
+
+/** Article Proofing Text */
 export const articleText = `A bizarre series of natural disasters has struck this once golden city. Eathquakes, tsunamis, volacanic eruptions, massive landslips, and sinkholes have caused Atlantis to sink over 10 kilometers into the Atlantic Ocean.Miraculously all individuals were able to be evacuated from the city a week prior, when authorities became deeply concered over the rising sea levels in the city due to the volcanic disaster in March this year.`;
 export const articleTitle = `BREAKING NEWS:`;
 export const articleSubTitle = `City of Atlantis deep in the Atlantic due to biggest natural disaster in history!`;
 export const articleAuthorLine = `Primary Reporter: Plato, Primary Journalist: Timaeus Critias, 360 BC`;
+
+/** Poster Proofing Text */
 export const posterTitle = 'Cat';
 export const posterText = 'Here come the claws';
 export const posterActorOne = 'Empurror Clawdius';
@@ -11,4 +15,5 @@ export const posterActorTwo = 'Queen Cleopatra';
 export const posterInfo =
   'Come to our furrtastic purrformace! Including treasured stars such as your favorite Empurror Clawdius and Queen Cleopatra! They will surely bat all doubts of purrfection aside, onto the floor, shattering it, leaving all you humans to clean up the mess.';
 export const posterTime = 'Book Meow!!!';
+
 export const maxHeightMultiplier = 100;
